+++ packages/pixel-craft/src/system/timer.ts
@@ -13,8 +13,8 @@ export class TimerSystem implements System {
    * The value will be between 0 and 1. If the value is 1, then the frame rate is at the target frame rate.
    */
   public deltaTime = 0;
+  public frameTime = 0;
   private lastFrame = performance.now();
-  private frameTime = 0;
 
   createSystem(): Promise<void> {
     this.lastFrame = performance.now();