+++ packages/tweening/src/index.ts
@@ -1,0 +1,2 @@
+export { easeInOutQuad, easeLinear } from './easing';
+export { Tween } from './tween';